Improve Elixir lexer based on Pygments
It appears that not all the rules from the Pygments Elixir lexer were
correctly ported over to Chroma. It was missing mostly the keywords and
declarations that are processed in the `get_tokens_unprocessed` function of the
Pygments Elixir lexer.

This commit adds the missing lines to the Elixir lexer and updates the
tests accordingly.