Fixed OpenEdgeABL String literals (#652)
When OpenEdge ABL runs on Windows, the tilde is used as an escape character instead of a backslash as in most common programming languages. On UNIX platforms, the backslash is also an escape character. But because we cannot know the running platform, it seems better to only use the tilde as a backslash character. (source: https://knowledgebase.progress.com/articles/Article/P27229)