Add Odin lexer (#802)
* Add Odin lexer

Adds the odin.xml rules for a lexer along with odin.actual and
odin.expected

Should handle the majority of Odin syntax well enough including nested comments.
However for now I have just left attributes as one token rather than sub-parsing
the @(x = y) to be multiple tokens

* Revert hlsl, natural, and vhs.expected back to master

These upated when I was generating odin.expected. I am
not sure if it is better to update these as well or leave it
so I figured I would leave for now

* Add the sized bool types and all the big/little endian int types

Odin has a number of other built in integer types that are not
in most other languages. Make sure those are included